feat: Add flag to give address of specific module
The `--module <addr>` option allows the address of a module to be given
to the following option. This currently only works for the `-f` option
to update a specific module instead of automatically selecting one.

As piControl already seems to implement updating specific modules and
not being limited to updating a module if just one module is connected
it can just be used in piTest, too.
More information about this can be found in `picontrol_ioctl(4)` or the
source code of piControl.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>