chore: release version 2.0.0
The semantic for the firmware update parameters has changed: While it
used to be allowed to omit a module address for the '-f' parameter (in
this case the first updateable module was chosen automatically) this is
not possible any more with the new firmware update implementation.

The exit codes as well as some of the error messages have changed as
well, and error messages are now output to stderr.

Hence update the major number since we are not backward compatible with
the new behaviour.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>