feat: add messages for firmware update
Add a warning message that informs the user that a firmware update is
starting and the system not to be shut down until the update is done.
Also add a hint to use the '--force' parameter if a firmware update failed
due to the firmware version check.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>
(cherry picked from commit eeda3f01edfcc211b5bb1cb9aad96afaeb4f099d)
Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>