feat(fwupdate): add spinner
When the firmware update is started there is no feedback from piTest
itself during the update. This is confusing as the caller doesn't know
for sure whether the update is running or not. Add a spinner to the
firmware update while the module is updating.

While the spinner doesn't show any progress information it still shows
that the firmware update is in progress. It's currently not possible to
implement a progress bar as the update itself is handled in piControl in
the kernel after the corresponding ioctl is called and no status
information is handed back to piTest for updating a progress bar.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>