fix: do not apply mask to active devices
For physical devices that have been configured via piCtory but not found
connected to the piBridge, the PICONTROL_NOT_CONNECTED bit is set in the
device ID.

In showDeviceList() the device ID is then ANDed with the
NOT_CONNECTED_MASK to mask out the bit and retrieve the original device
ID.

However the NOT_CONNECTED_MASK is 0x7fff and ANDing it with the ID
0x8000 (Virtual Device 32 Byte) results in the incorrect ID 0x0000.

Avoid this issue by not applying the mask to active devices. This
approach works, because the PICONTROL_NOT_CONNECTED bit is only set for
inactive devices and furthermore virtual devices are always 'active'.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>
(cherry picked from commit 70d749416debaa390acd69be81cc0065f02a82d8)
Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>