feat: Remove redundant error messages in fw update
The current implementation issues 3 redundant error messages, when a
firmware update fails:

$ src/piTest --module 31 --force -f
Are you sure you want to update the firmware of a RevPi module? (y/N) y
Updating Firmware (forced)!
This can take a while. Do not switch off the system!
Failed to update firmware of module with address 31: Operation not permitted
failed to update firmware: Operation not permitted
error when updating firmware: Operation not permitted

Fix this and use only the one in piControlIf.c.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>