fix: correct parameter of function writeData
Function writeData() only accepts 4-byte values, so turn the data type
for parameter 'i32uValue' from an unsigned long which may be an 8-byte
value on 64-bit systems into an unsigned int which is a 4-byte value on
both 32 and 64-bit systems.

Also for option case 's' use an unsigned int instead of an unsigned long
to store the value parsed from a string given by the user.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>